refactor(EntitiesList): Refactor EntitiesList Component. Closes #52
